+++ geoopt/optim/radam.py
@@ -4,6 +4,7 @@ from .mixin import OptimMixin
 from .tracing import create_traced_update
 from ..tensor import ManifoldParameter, ManifoldTensor
 from ..manifolds import Euclidean
+from ..utils import copy_or_set_
 
 
 class RiemannianAdam(OptimMixin, torch.optim.Adam):
@@ -189,7 +190,8 @@ class RiemannianAdam(OptimMixin, torch.optim.Adam):
         new_point, exp_avg_new = manifold.retr_transp(
             point, exp_avg, u=direction, t=-step_size
         )
-        point.set_(new_point)
+        # use copy only for user facing point
+        copy_or_set_(point, new_point)
         exp_avg.set_(exp_avg_new)
 
     def stabilize_group(self, group):
@@ -202,7 +204,7 @@ class RiemannianAdam(OptimMixin, torch.optim.Adam):
                     continue
                 manifold = p.manifold
                 exp_avg = state["exp_avg"]
-                p.set_(manifold.projx(p))
+                copy_or_set_(p, manifold.projx(p))
                 exp_avg.set_(manifold.proju(p, exp_avg))
 
     def _sanitize_group(self, group):