+++ geoopt/optim/rsgd.py
@@ -3,7 +3,7 @@ from ..manifolds import Euclidean
 from ..tensor import ManifoldParameter, ManifoldTensor
 from .mixin import OptimMixin
 from .tracing import create_traced_update
-
+from ..utils import copy_or_set_
 
 __all__ = ["RiemannianSGD"]
 
@@ -168,10 +168,11 @@ class RiemannianSGD(OptimMixin, torch.optim.Optimizer):
                 point, momentum_buffer, u=grad, t=-lr
             )
             momentum_buffer.set_(new_momentum_buffer)
-            point.set_(new_point)
+            # use copy only for user facing point
+            copy_or_set_(point, new_point)
         else:
             new_point = manifold.retr(point, grad, -lr)
-            point.set_(new_point)
+            copy_or_set_(point, new_point)
 
     def stabilize_group(self, group):
         with torch.no_grad():
@@ -180,7 +181,7 @@ class RiemannianSGD(OptimMixin, torch.optim.Optimizer):
                     continue
                 manifold = p.manifold
                 momentum = group["momentum"]
-                p.set_(manifold.projx(p))
+                copy_or_set_(p, manifold.projx(p))
                 if momentum > 0:
                     param_state = self.state[p]
                     if not param_state:  # due to None grads